refactor: split four files back under the 2000-line cap
#9505 took child_process/reactor.rs to 2283 and fs/stream.rs to 2140, #9507
took dynamic_dispatch.rs to 2029, #9508 took date.rs to 2067. Each split
follows its file's existing sibling convention: date/tests.rs,
property_get/dispatch_receiver_class.rs, fs/stream/options_init.rs, and
reactor/{kill,stdin_drain}.rs as child modules reaching parent privates.
cp_live_kill keeps pub(crate) for emitter.rs's cross-module call.
